requirements.txt
Django==1.11
pytz==2018.7
wheel==0.24.0
django-registration-redux==2.4
